Enhance order management and introduce new order calculation functions
- Refactored `update_order_after_fill` to simplify return statements.
- Added `should_reward_keeper` function to determine if a keeper should be rewarded based on user and filler authorities.
- Introduced `calculate_quote_asset_amount_for_maker_order` to compute quote asset amounts based on base asset amounts and prices.
- Implemented `should_expire_order` and `should_cancel_reduce_only_order` functions for better order lifecycle management.
- Added unit tests for new functions to ensure correct behavior and edge case handling.
- Updated `Order` struct methods for improved logic and clarity.
- Enhanced `matching` module with new fill calculation logic for matched orders.

+++ programs/mini-drift/src/state/fill_mode.rs
@@ -40,7 +40,7 @@ impl FillMode {
                         .ok_or(ErrorCode::MathError)?;
                     let pretended_slots = order
                         .slot
-                        .checked_add(calc_slot as u64)
+                        .checked_add(calc_slot)
                         .ok_or(ErrorCode::MathError)?;
                     let auction_price = calculate_auction_price(order, pretended_slots)?;
                     Ok(Some(auction_price))
